fixed tests again
